Added tasks to the TODO list.
Converted the makefile style transformation of a computation into
a command into a more generalized external command form of the
computations.  Condensed functionality from driver-commands.lisp and
makefile-backend.lisp into external-commands.lisp. This is because at
this time both the makefile backend and a new run-program back end will
use them.

Added a run-program based backend called "simple-build" which will execute
the commands necessary to build the dependencies in the graph. At this
time it is a serial execution of the dependencies. The functionality is
not yet complete and currently broken as of this checkin.

Fixed up the xcvb.asd file to denote the new dependencies.