Made more progress on the run-program backend. It now actually functions and runs the computations to produce the fasl/cfasl files! I don't want to up the revision number since the function SIMPLE-BUILD still needs cleaning up in relation to an apparently meaningless :output keyword i nthe function's lambda-list, among other things.
Fixed ALL-NEWER-PATHNAMES-P to additionally understand that when
output files do not exist, then the output file set can't possibly
be newer than the input set. We return NIL in this case which forces
the regeneration of the ouput files.

Added some comments for things we need to check later in the code.

Some minor code formatting changes.