Cosmetic indention change of a piece of code.
Converted more ERROR calls to USER-ERROR and fixed up the message to
be more clear.

When the simple-dispatcher constructs a dispatch set of forms, the
name of the dispatcher is now embedded into it. This makes it
much easier to find _which_ simple dispatcher interpreter might be
missing how to evaluate a form.